# apps/command/openapi.json
{
  "components": {
    "schemas": {
...
      "GeoPoint": {
        "description": "A WGS84 point.",
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "lon",
          "lat"
        ],
        "title": "GeoPoint",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Outlook": {
        "description": "Today's next three hours of street depth from live rain, labelled as what it is.",
        "properties": {
          "baseline_removed": {
            "$ref": "#/components/schemas/OutlookBaseline"
          },
          "blockage": {
            "$ref": "#/components/schemas/OutlookBlockage"
          },
          "cached": {
            "description": "True when this answer was computed for an earlier request.",
            "title": "Cached",
            "type": "boolean"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "compute_ms": {
            "title": "Compute Ms",
            "type": "integer"
          },
          "expired": {
            "description": "True once `valid_to` has passed (a stale weather copy).",
            "title": "Expired",
            "type": "boolean"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "members": {
            "title": "Members",
            "type": "integer"
          },
          "method": {
            "default": "reduced-order emulator (Flash-lite) on Open-Meteo NWP rain, AOI-uniform; not a radar nowcast",
            "title": "Method",
            "type": "string"
          },
          "mode": {
            "const": "outlook",
            "default": "outlook",
            "title": "Mode",
            "type": "string"
          },
          "n_segments_over_1cm": {
            "title": "N Segments Over 1Cm",
            "type": "integer"
          },
          "n_steps": {
            "title": "N Steps",
            "type": "integer"
          },
          "no_rain_response": {
            "description": "Streets the emulator has no rain response for (never wet in training); they read 0 cm here whatever the rain.",
            "title": "No Rain Response",
            "type": "integer"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "rain_mm_h": {
            "description": "AOI-uniform rain rate per step, mm/h.",
            "items": {
              "type": "number"
            },
            "title": "Rain Mm H",
            "type": "array"
          },
          "rain_total_mm": {
            "title": "Rain Total Mm",
            "type": "number"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Always null: an outlook is not a run and is not registered.",
            "title": "Run Id"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/OutlookSegment"
            },
            "title": "Segments",
            "type": "array"
          },
          "skill": {
            "$ref": "#/components/schemas/OutlookSkill"
          },
          "source": {
            "$ref": "#/components/schemas/OutlookSource"
          },
          "steps_min": {
            "default": 5,
            "title": "Steps Min",
            "type": "integer"
          },
          "summary": {
            "$ref": "#/components/schemas/OutlookSummary"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "unmatched_streets": {
            "description": "The emulator's streets whose id the city's current street table does not carry: forecast, but with no name and no geometry in today's street layer.",
            "title": "Unmatched Streets",
            "type": "integer"
          },
          "valid_from": {
            "format": "date-time",
            "title": "Valid From",
            "type": "string"
          },
          "valid_to": {
            "format": "date-time",
            "title": "Valid To",
            "type": "string"
          },
          "valid_ts": {
            "description": "End of each 5-minute step.",
            "items": {
              "format": "date-time",
              "type": "string"
            },
            "title": "Valid Ts",
            "type": "array"
          }
        },
        "required": [
          "city",
          "issued_at",
          "valid_from",
          "valid_to",
          "expired",
          "n_steps",
          "valid_ts",
          "rain_mm_h",
          "rain_total_mm",
          "members",
          "source",
          "segments",
          "n_segments_over_1cm",
          "truncated",
          "summary",
          "baseline_removed",
          "blockage",
          "skill",
          "no_rain_response",
          "unmatched_streets",
          "notes",
          "compute_ms",
          "cached"
        ],
        "title": "Outlook",
        "type": "object"
      },
      "OutlookBaseline": {
        "description": "The base state taken out: what the emulator would have drawn with no rain at all.",
        "properties": {
          "max_cm": {
            "title": "Max Cm",
            "type": "number"
          },
          "n_ge_15": {
            "title": "N Ge 15",
            "type": "integer"
          },
          "n_ge_30": {
            "title": "N Ge 30",
            "type": "integer"
          },
          "n_ge_5": {
            "title": "N Ge 5",
            "type": "integer"
          }
        },
        "required": [
          "n_ge_5",
          "n_ge_15",
          "n_ge_30",
          "max_cm"
        ],
        "title": "OutlookBaseline",
        "type": "object"
      },
...
      "Weather": {
        "description": "Current conditions and the next four hours, with its provenance and its age.",
        "properties": {
          "age_s": {
            "description": "Seconds since `fetched_at`, computed per request.",
            "title": "Age S",
            "type": "number"
          },
          "attribution": {
            "default": "Weather data by Open-Meteo.com (CC BY 4.0)",
            "title": "Attribution",
            "type": "string"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "current": {
            "$ref": "#/components/schemas/WeatherNow"
          },
          "elevation_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "fetched_at": {
            "description": "When this copy was retrieved from Open-Meteo.",
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "grid_offset_km": {
            "description": "Distance between the two, kilometres.",
            "title": "Grid Offset Km",
            "type": "number"
          },
          "grid_point": {
            "$ref": "#/components/schemas/GeoPoint",
            "description": "The centre of the model cell that answered."
          },
          "hourly": {
            "items": {
              "$ref": "#/components/schemas/WeatherStep"
            },
            "title": "Hourly",
            "type": "array"
          },
          "licence": {
            "default": "CC BY 4.0",
            "title": "Licence",
            "type": "string"
          },
          "licence_url": {
            "default": "https://creativecommons.org/licenses/by/4.0/",
            "title": "Licence Url",
            "type": "string"
          },
          "minutely_15": {
            "description": "Open-Meteo's 15-minute precipitation for the next four hours. Outside Europe and North America it is interpolated from hourly models; empty in copies fetched before it was asked for.",
            "items": {
              "$ref": "#/components/schemas/WeatherStep"
            },
            "title": "Minutely 15",
            "type": "array"
          },
          "notes": {
            "description": "Why this copy is what it is; printed by the UI.",
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "point": {
            "$ref": "#/components/schemas/GeoPoint",
            "description": "The city AOI centre this was asked for."
          },
          "source": {
            "default": "open-meteo",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "default": "https://open-meteo.com/",
            "title": "Source Url",
            "type": "string"
          },
          "stale": {
            "description": "True once `age_s` exceeds the 900 s cache window.",
            "title": "Stale",
            "type": "boolean"
          },
          "ttl_s": {
            "default": 900.0,
            "title": "Ttl S",
            "type": "number"
          }
        },
        "required": [
          "city",
          "point",
          "grid_point",
          "grid_offset_km",
          "current",
          "hourly",
          "fetched_at",
          "age_s",
          "stale"
        ],
        "title": "Weather",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Team VIT, SIH 2026 PS SIH26085"
    },
    "description": "Street-level urban flood nowcasting digital twin (SIH 2026, PS SIH26085, MoES). Every response carries `run_id` and `valid_ts` where a run is involved; errors use `{ error: { code, message, run_id } }`. Endpoints answering 501 are contract-only until their engine's phase lands.",
    "title": "VARUNA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/outlook": {
      "get": {
        "description": "A labelled live outlook: Open-Meteo NWP rain, AOI-uniform, through Flash-lite at 50 members.\n\nDepth is reported **above** the emulator's stored base state, never with it: that base state\nis a training window's tide and upstream runoff and would draw a flood on a dry day. Cached\nwith the weather copy it came from (15 minutes); a stale copy is served with its age and its\nown window, never shifted onto the present.",
        "operationId": "outlook_v1_outlook_get",
        "parameters": [
          {
            "description": "City slug; Mumbai is the only one fitted.",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "mumbai",
              "description": "City slug; Mumbai is the only one fitted.",
              "title": "City",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Outlook"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "No emulator is fitted to this city's streets (Chennai)"
          },
          "503": {
            "description": "No weather copy, no fitted emulator, the city is not built, or the emulator was fitted to an older build of it"
          }
        },
        "summary": "Today's next 3 h of street depth from live Open-Meteo rain, through the emulator",
        "tags": [
          "weather"
        ]
      }
    },
...
    "/v1/route": {
      "post": {
        "description": "Plan one trip.\n\nBody: ``{origin, destination, depart_at?, profile?, risk_tolerance?, run_id?, spread?,\ntrip_id?, explain?}``.\n\nReturns both routes, because the comparison is the product: a dispatcher who only sees the\nsafe route has no way to judge whether the detour was worth it.\n\n``spread`` (default true) adds ``corridors[]`` - up to three safe roads with the share of\ntraffic the policy gives each, and the one this request is assigned to. ``trip_id`` makes\nthat assignment stable for a trip, so a reader who reloads is not sent somewhere else; it is\nthe client's own random id and nothing is stored against it. ``explain`` (default true) adds\n``reasons[]``, which is structured data - the frontend writes the sentences (TECH_SPEC 3.2).",
        "operationId": "route_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route V1 Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route around the forecast water, beside what a naive router does",
        "tags": [
          "route"
        ]
      }
    },
...
    "/v1/weather": {
      "get": {
        "description": "Live weather for the city's AOI centre, cached for fifteen minutes.\n\nDegrades rather than fails: an upstream error, a timeout or `VARUNA_OFFLINE=1` serves the last\ngood copy with its true age and `stale: true`. Only when no copy has ever been fetched does\nthis answer 503, with the reason named.",
        "operationId": "weather_v1_weather_get",
        "parameters": [
          {
            "description": "City slug, e.g. mumbai or chennai.",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "mumbai",
              "description": "City slug, e.g. mumbai or chennai.",
              "title": "City",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current conditions and the next four hours from Open-Meteo",
        "tags": [
          "weather"
        ]
      }
    },
...
  },
...
}